Added word banks for each stocktype to give real(ish) names to stock in a water treatment context
StockInitializationService can now support generating these names, as well as a random ratio of stock type items (still in the bounds of 196 items). The "distance" is pretty good although there is some spill over between stockrooms of the same stock types. No point fixing this as the next rule to be implemented is to place stock that get frequently picked in the same orders together.